Added a setup.py, changed file names
