+++ .gitignore
@@ -1,2 +1,3 @@
 .venv
 *.pyc
+/keymanager.egg-info/
