+++ setup.py
@@ -1,0 +1,34 @@
+#!/usr/bin/env python
+
+try:
+    from setuptools import setup
+except ImportError:
+    from distutils.core import setup
+
+setup(
+    name='keymanager',
+    version='1.0.2',
+    author='farridav',
+    author_email='info@davidfarrington.co.uk',
+    entry_points={
+        'console_scripts': [
+            'keymanager = keymanager:keymanager_main',
+        ],
+    },
+    url='https://github.com/farridav/keymanager.git',
+    license='MIT',
+    description='SSH Key manager, powered by fabric',
+    classifiers=[
+        'Development Status :: 3 - Alpha',
+        'Environment :: Web Environment',
+        'Intended Audience :: Developers',
+        'Framework :: Fabric',
+        'License :: OSI Approved :: MIT License',
+        'Natural Language :: English',
+        'Operating System :: OS Independent',
+        'Topic :: Internet :: WWW/HTTP',
+    ],
+    install_requires=[
+        "Fabric>=1.9.0",
+    ],
+)
